# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "CloudLens API",
    "description": "Azure FinOps Managed Service \u2014 cost intelligence API",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/forecast/{tenant_id}": {
      "get": {
        "tags": [
          "forecast"
        ],
        "summary": "Get Spend Forecast",
        "description": "Baseline spend forecast (Holt-Winters, weekly seasonality, with backtest MAPE).",
        "operationId": "get_spend_forecast_api_v1_forecast__tenant_id__get",
        "parameters": [
          {
            "name": "tenant_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Tenant Id"
            }
          },
          {
            "name": "horizon_days",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 90,
              "minimum": 7,
              "default": 30,
              "title": "Horizon Days"
            }
          },
          {
            "name": "history_days",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 90,
              "minimum": 14,
              "default": 90,
              "title": "History Days"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SpendForecastResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/forecast/{tenant_id}/cost-of-inaction": {
      "get": {
        "tags": [
          "forecast"
        ],
        "summary": "Get Cost Of Inaction",
        "description": "Dual-trajectory forecast: do-nothing vs. if-you-act, with cost-of-inaction.",
        "operationId": "get_cost_of_inaction_api_v1_forecast__tenant_id__cost_of_inaction_get",
        "parameters": [
          {
            "name": "tenant_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Tenant Id"
            }
          },
          {
            "name": "horizon_days",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 90,
              "minimum": 7,
              "default": 30,
              "title": "Horizon Days"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TrajectoryResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/forecast/{tenant_id}/roadmap": {
      "get": {
        "tags": [
          "forecast"
        ],
        "summary": "Get Remediation Roadmap",
        "description": "ROI-ordered, phased remediation roadmap with run-rate bending down per phase.",
        "operationId": "get_remediation_roadmap_api_v1_forecast__tenant_id__roadmap_get",
        "parameters": [
          {
            "name": "tenant_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Tenant Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoadmapResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/forecast/{tenant_id}/budget-breach": {
      "get": {
        "tags": [
          "forecast"
        ],
        "summary": "Get Budget Breach",
        "description": "Predict the date cumulative spend breaches the budget, do-nothing vs. if-actioned.",
        "operationId": "get_budget_breach_api_v1_forecast__tenant_id__budget_breach_get",
        "parameters": [
          {
            "name": "tenant_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Tenant Id"
            }
          },
          {
            "name": "monthly_budget",
            "in": "query",
            "required": true,
            "schema": {
              "type": "number",
              "exclusiveMinimum": 0,
              "description": "Monthly budget in EUR",
              "title": "Monthly Budget"
            },
            "description": "Monthly budget in EUR"
          },
          {
            "name": "horizon_days",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 90,
              "minimum": 7,
              "default": 60,
              "title": "Horizon Days"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BudgetBreachResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BudgetBreachResponse": {
        "properties": {
          "tenant_id": {
            "type": "string",
            "title": "Tenant Id"
          },
          "monthly_budget_eur": {
            "type": "number",
            "title": "Monthly Budget Eur"
          },
          "breach_date_baseline": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Breach Date Baseline"
          },
          "breach_date_optimized": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Breach Date Optimized"
          },
          "safe_if_actioned": {
            "type": "boolean",
            "title": "Safe If Actioned"
          },
          "notes": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Notes"
          }
        },
        "type": "object",
        "required": [
          "tenant_id",
          "monthly_budget_eur",
          "safe_if_actioned"
        ],
        "title": "BudgetBreachResponse"
      },
...
      "ForecastPointModel": {
        "properties": {
          "day": {
            "type": "string",
            "title": "Day"
          },
          "value": {
            "type": "number",
            "title": "Value"
          },
          "lower": {
            "type": "number",
            "title": "Lower"
          },
          "upper": {
            "type": "number",
            "title": "Upper"
          }
        },
        "type": "object",
        "required": [
          "day",
          "value",
          "lower",
          "upper"
        ],
        "title": "ForecastPointModel"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "RoadmapPhaseModel": {
        "properties": {
          "phase": {
            "type": "integer",
            "title": "Phase"
          },
          "label": {
            "type": "string",
            "title": "Label"
          },
          "items": {
            "type": "integer",
            "title": "Items"
          },
          "monthly_saving_eur": {
            "type": "number",
            "title": "Monthly Saving Eur"
          },
          "cumulative_monthly_saving_eur": {
            "type": "number",
            "title": "Cumulative Monthly Saving Eur"
          },
          "target_run_rate_eur": {
            "type": "number",
            "title": "Target Run Rate Eur"
          },
          "eta_days": {
            "type": "integer",
            "title": "Eta Days"
          }
        },
        "type": "object",
        "required": [
          "phase",
          "label",
          "items",
          "monthly_saving_eur",
          "cumulative_monthly_saving_eur",
          "target_run_rate_eur",
          "eta_days"
        ],
        "title": "RoadmapPhaseModel"
      },
      "RoadmapResponse": {
        "properties": {
          "tenant_id": {
            "type": "string",
            "title": "Tenant Id"
          },
          "current_run_rate_eur": {
            "type": "number",
            "title": "Current Run Rate Eur"
          },
          "optimized_run_rate_eur": {
            "type": "number",
            "title": "Optimized Run Rate Eur"
          },
          "total_monthly_saving_eur": {
            "type": "number",
            "title": "Total Monthly Saving Eur"
          },
          "phases": {
            "items": {
              "$ref": "#/components/schemas/RoadmapPhaseModel"
            },
            "type": "array",
            "title": "Phases"
          }
        },
        "type": "object",
        "required": [
          "tenant_id",
          "current_run_rate_eur",
          "optimized_run_rate_eur",
          "total_monthly_saving_eur"
        ],
        "title": "RoadmapResponse"
      },
...
      "SpendForecastResponse": {
        "properties": {
          "tenant_id": {
            "type": "string",
            "title": "Tenant Id"
          },
          "method": {
            "type": "string",
            "title": "Method"
          },
          "horizon_days": {
            "type": "integer",
            "title": "Horizon Days"
          },
          "history_days": {
            "type": "integer",
            "title": "History Days"
          },
          "mape": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mape",
            "description": "Backtest mean abs % error"
          },
          "confidence": {
            "type": "string",
            "title": "Confidence"
          },
          "month_end_projection": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Month End Projection"
          },
          "points": {
            "items": {
              "$ref": "#/components/schemas/ForecastPointModel"
            },
            "type": "array",
            "title": "Points"
          },
          "notes": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Notes"
          }
        },
        "type": "object",
        "required": [
          "tenant_id",
          "method",
          "horizon_days",
          "history_days",
          "confidence"
        ],
        "title": "SpendForecastResponse"
      },
...
      "TrajectoryResponse": {
        "properties": {
          "tenant_id": {
            "type": "string",
            "title": "Tenant Id"
          },
          "horizon_days": {
            "type": "integer",
            "title": "Horizon Days"
          },
          "daily_waste_burn_eur": {
            "type": "number",
            "title": "Daily Waste Burn Eur"
          },
          "cumulative_inaction_eur": {
            "type": "number",
            "title": "Cumulative Inaction Eur"
          },
          "monthly_recoverable_eur": {
            "type": "number",
            "title": "Monthly Recoverable Eur"
          },
          "annual_recoverable_eur": {
            "type": "number",
            "title": "Annual Recoverable Eur"
          },
          "baseline": {
            "items": {
              "$ref": "#/components/schemas/ForecastPointModel"
            },
            "type": "array",
            "title": "Baseline"
          },
          "optimized": {
            "items": {
              "$ref": "#/components/schemas/ForecastPointModel"
            },
            "type": "array",
            "title": "Optimized"
          },
          "notes": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Notes"
          }
        },
        "type": "object",
        "required": [
          "tenant_id",
          "horizon_days",
          "daily_waste_burn_eur",
          "cumulative_inaction_eur",
          "monthly_recoverable_eur",
          "annual_recoverable_eur"
        ],
        "title": "TrajectoryResponse"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}